virt-controller, template: move CPU labels / feature extraction funcs to the renderer
